Moved Listener to core.py
