+++ ucp/public_api.py
@@ -201,35 +201,3 @@ def reset():
             for o in gc.get_referrers(weakref_ctx()):
                 msg += "\n  %s" % o
             raise exceptions.UCXError(msg)
-
-
-class Listener:
-    """A handle to the listening service started by `create_listener()`
-
-    The listening continues as long as this object exist or `.close()` is called.
-    Please use `create_listener()` to create an Listener.
-    """
-
-    def __init__(self, backend):
-        self._b = backend
-        self._closed = False
-
-    def __del__(self):
-        if not self.closed():
-            self.close()
-
-    def closed(self):
-        """Is the listener closed?"""
-        return self._closed
-
-    @property
-    def port(self):
-        """The network point listening on"""
-        return self._b.port
-
-    def close(self):
-        """Closing the listener"""
-        if not self._closed:
-            self._b.abort()
-            self._closed = True
-            self._b = None
